requirements.txt
# Core ML
scikit-learn==1.4.2
pandas==2.2.1
numpy==1.26.4
scipy==1.13.0

# MLflow
mlflow==2.12.1
boto3==1.34.84
psycopg2-binary==2.9.9

# Airflow (installed separately via docker)
apache-airflow==2.8.1

# FastAPI serving
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
httpx==0.27.0

# Monitoring
evidently==0.4.22

# Hyperparameter tuning
optuna==3.6.1

# Utilities
python-dotenv==1.0.1
requests==2.31.0
click==8.1.7
PyYAML==6.0.1

# Testing
pytest==8.1.1
pytest-cov==5.0.0
pytest-asyncio==0.23.6

# Linting
ruff==0.4.2

# Notifications
slack-sdk==3.27.1
